custom-elements/instantiating-custom-elements/changing-is-attribute.html

<!DOCTYPE html>
<html>
<head>
<title>Changing IS attribute of the custom element after element is instantiated</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="After a custom element is instantiated, changing the value of the is attribute must not affect this element's custom element type">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#instantiating-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(tagName) {
        if (HTML5_DOCUMENT_ELEMENTS.indexOf(tagName) !== -1) {
            return;
        }
        var name = 'x-' + tagName;
        doc.registerElement(name);
        if (HTML5_TABLE_ELEMENTS.indexOf(tagName) !== -1) {
            doc.body.innerHTML =
                '<table>' +
                '<' + tagName + ' id="custom-element" is="' + name + '"></' + tagName + '>' +
                '</table>';
        } else {
            doc.body.innerHTML =
                '<' + tagName + ' id="custom-element" is="' + name + '"></' + tagName + '>';
        }
        var customElement = doc.querySelector('#custom-element');
        var obj = doc.createElement(tagName);
        assert_class_string(customElement, obj.constructor.name,
            'Custom element must be ' + obj.constructor.name);

        var name2 = 'x-a-' + tagName;
        doc.registerElement(name2);
        customElement.setAttribute('is', name2);
        assert_class_string(customElement, obj.constructor.name,
            'Custom element must not be changed and should be ' + obj.constructor.name);
    });
}, 'Changing IS attribute of the custom element after element is instatiated');


test(function() {
    var doc = newHTMLDocument();
    var localName = 'x-b-a';
    doc.registerElement(localName);
    doc.body.innerHTML = '<a id="custom-element" is="' + localName + '"></a>';
    var customElement = doc.querySelector('#custom-element');
    var obj = doc.createElement('a');

    HTML5_ELEMENTS.forEach(function(tagName) {
        var name = 'x-c-' + tagName;
        var htmlElement = doc.createElement(tagName);
        var proto = Object.create(htmlElement.constructor.prototype);
        doc.registerElement(name, {prototype: proto, extends: tagName});
        customElement.setAttribute('is', name);
        assert_class_string(customElement, obj.constructor.name,
            'Custom element must not be changed and should be ' + obj.constructor.name);
    });
}, 'Changing IS attribute of the custom element after element is instatiated. Test all possible values ' +
    'of extensions of HTML elements as new IS value');
</script>
</body>
</html>
